moved all components to appropriate packages
