+++ app/src/main/java/com/doplgangr/secrecy/activities/MainActivity.java
@@ -41,12 +41,12 @@ import android.widget.TextView;
 import com.doplgangr.secrecy.Config;
 import com.doplgangr.secrecy.CustomApp;
 import com.doplgangr.secrecy.R;
-import com.doplgangr.secrecy.Util;
+import com.doplgangr.secrecy.utils.Util;
 import com.doplgangr.secrecy.filesystem.Storage;
 import com.doplgangr.secrecy.filesystem.encryption.VaultHolder;
-import com.doplgangr.secrecy.premium.PremiumFragment;
-import com.doplgangr.secrecy.settings.SettingsFragment;
-import com.doplgangr.secrecy.updatemanager.UpdateManager;
+import com.doplgangr.secrecy.fragments.PremiumFragment;
+import com.doplgangr.secrecy.fragments.SettingsFragment;
+import com.doplgangr.secrecy.fragments.UpdateManager;
 import com.doplgangr.secrecy.adapters.VaultsListFragment;
 import com.doplgangr.secrecy.views.navdrawer.DrawerLayout;
 import com.doplgangr.secrecy.views.navdrawer.NavItem;
