+++ app/src/main/java/com/doplgangr/secrecy/fragments/FileViewer.java
@@ -37,7 +37,7 @@ import android.widget.Toast;
 import com.doplgangr.secrecy.Config;
 import com.doplgangr.secrecy.CustomApp;
 import com.doplgangr.secrecy.R;
-import com.doplgangr.secrecy.Util;
+import com.doplgangr.secrecy.utils.Util;
 import com.doplgangr.secrecy.activities.FilesActivity;
 import com.doplgangr.secrecy.filesystem.CryptStateListener;
 import com.doplgangr.secrecy.filesystem.OurFileProvider;
@@ -45,7 +45,6 @@ import com.doplgangr.secrecy.filesystem.Storage;
 import com.doplgangr.secrecy.filesystem.encryption.Vault;
 import com.doplgangr.secrecy.filesystem.files.EncryptedFile;
 import com.doplgangr.secrecy.filesystem.files.SecrecyFile;
-import com.doplgangr.secrecy.fragments.FilesListFragment;
 import com.doplgangr.secrecy.jobs.AddFileJob;
 
 import java.io.File;
