+++ app/src/main/java/com/doplgangr/secrecy/fragments/SettingsFragment.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package com.doplgangr.secrecy.settings;
+package com.doplgangr.secrecy.fragments;
 
 import android.app.Activity;
 import android.app.AlertDialog;
@@ -49,9 +49,8 @@ import android.widget.Toast;
 import com.doplgangr.secrecy.Config;
 import com.doplgangr.secrecy.CustomApp;
 import com.doplgangr.secrecy.R;
-import com.doplgangr.secrecy.Util;
+import com.doplgangr.secrecy.utils.Util;
 import com.doplgangr.secrecy.filesystem.Storage;
-import com.doplgangr.secrecy.premium.PremiumFragment;
 import com.doplgangr.secrecy.premium.PremiumStateHelper;
 import com.doplgangr.secrecy.premium.StealthMode;
 import com.doplgangr.secrecy.adapters.VaultsListFragment;
