+++ app/src/main/java/com/doplgangr/secrecy/jobs/AddFileJob.java
@@ -9,7 +9,7 @@ import com.doplgangr.secrecy.events.NewFileEvent;
 import com.doplgangr.secrecy.filesystem.encryption.Vault;
 import com.doplgangr.secrecy.filesystem.files.EncryptedFile;
 import com.doplgangr.secrecy.filesystem.Storage;
-import com.doplgangr.secrecy.Util;
+import com.doplgangr.secrecy.utils.Util;
 import com.path.android.jobqueue.Job;
 import com.path.android.jobqueue.Params;
