+++ app/src/main/java/com/doplgangr/secrecy/jobs/DeleteFileJob.java
@@ -8,7 +8,7 @@ import android.provider.DocumentsContract;
 
 import com.doplgangr.secrecy.CustomApp;
 import com.doplgangr.secrecy.filesystem.Storage;
-import com.doplgangr.secrecy.Util;
+import com.doplgangr.secrecy.utils.Util;
 import com.path.android.jobqueue.Job;
 import com.path.android.jobqueue.Params;
