+++ app/src/main/java/com/doplgangr/secrecy/jobs/ThumbnailLoadJob.java
@@ -6,7 +6,7 @@ import android.widget.ImageView;
 import com.doplgangr.secrecy.events.ThumbLoadDoneEvent;
 import com.doplgangr.secrecy.exceptions.SecrecyFileException;
 import com.doplgangr.secrecy.filesystem.files.EncryptedFile;
-import com.doplgangr.secrecy.Util;
+import com.doplgangr.secrecy.utils.Util;
 import com.path.android.jobqueue.Job;
 import com.path.android.jobqueue.Params;
